Splits CODE_MISSING_PARAMETER_TYPE_HINT into two
For some reason constant CODE_MISSING_PARAMETER_TYPE_HINT is used for two different situations – one is for missing the doc and the hint in the same time, the other one is for suggesting using hint, because it should be possible (the message says).

In many cases, it’s OK and I should always use hint, but in my case I just inherited a method from a parent, who isn’t mine but it’s from a framework. In this case I’m not able to use hint, because it breaks the PHP rule about the “parameters in all inherited methods must be the same type”. Funny is that setting the return type declarations isn’t a problem. 

So in this PR I’ve split this constant into two separated and thanks to that I’m able to use exclude in my own ruleset and still keep using the annotation for missing hint and docs in the same time.